+++ src/main/java/duke/commands/Command.java
@@ -1,9 +1,9 @@
-package commands;
+package duke.commands;
 
-import task.TaskList;
-import ui.Ui;
+import duke.task.TaskList;
+import duke.ui.Ui;
 import duke.DukeException;
-import storage.Storage;
+import duke.storage.Storage;
 
 public abstract class Command {
     public boolean isExit = false;
