+++ src/main/java/duke/commands/ExitCommand.java
@@ -1,9 +1,9 @@
-package commands;
+package duke.commands;
 
-import task.TaskList;
-import ui.Ui;
+import duke.task.TaskList;
+import duke.ui.Ui;
 import duke.DukeException;
-import storage.Storage;
+import duke.storage.Storage;
 
 public class ExitCommand  extends Command {
     public ExitCommand() {
