+++ src/main/java/duke/commands/ListCommand.java
@@ -1,9 +1,9 @@
-package commands;
+package duke.commands;
 
-import task.TaskList;
-import ui.Ui;
+import duke.task.TaskList;
+import duke.ui.Ui;
 import duke.DukeException;
-import storage.Storage;
+import duke.storage.Storage;
 
 public class ListCommand extends Command {
     public ListCommand() {}
