+++ src/main/java/duke/task/TaskList.java
@@ -1,8 +1,7 @@
-package task;
+package duke.task;
 import java.util.ArrayList;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
-import java.util.stream.Stream;
 public class TaskList {
     private ArrayList<Task> list;
